refactor: extract BaseLoan ABC from Loan and BillingCycleLoan
Move 22 identical methods (balance properties, pay_installment,
schedule queries, fine tracking, Warp hooks, settlement/installment
views) into a shared BaseLoan ABC.  Both Loan and BillingCycleLoan
now extend BaseLoan and only implement the abstract hooks that differ:
_compute_state, _accrued_interest_components, _build_initial_cashflow,
settlement_balance, and record_payment.

Pure structural refactoring -- zero behavioral changes.  All 5692
existing tests pass unchanged.
