+++ money_warp/billing_cycle_loan/billing_cycle_loan.py
@@ -1,23 +1,20 @@
 """BillingCycleLoan -- fixed amortization with billing-cycle payment timing."""
 
-import warnings
 from datetime import date, datetime, tzinfo
-from typing import Dict, List, Optional, Type, Union
+from typing import Dict, List, Optional, Tuple, Type, Union
 from zoneinfo import ZoneInfo
 
+from ..base_loan import BaseLoan
 from ..billing_cycle import BaseBillingCycle
 from ..cash_flow import CashFlow, CashFlowItem, CashFlowType
 from ..engines import (
     InterestCalculator,
     LoanState,
     MoraStrategy,
-    apply_tolerance_adjustment,
-    build_installments,
     covered_due_date_count,
-    is_payment_late,
 )
-from ..models import BillingCycleLoanStatement, Installment, Settlement
-from ..scheduler import BaseScheduler, PaymentSchedule, PaymentScheduleEntry, PriceScheduler
+from ..models import BillingCycleLoanStatement, Settlement
+from ..scheduler import BaseScheduler, PriceScheduler
 from ..time_context import TimeContext
 from ..types.interest_rate import InterestRate
 from ..types.money import Money
@@ -27,7 +24,7 @@ from .engines import build_statements, compute_state, resolve_mora_rate
 from .mora_rate_resolver import MoraRateResolver
 
 
-class BillingCycleLoan:
+class BillingCycleLoan(BaseLoan):
     """Loan with fixed amortization and credit-card-like billing cycles.
 
     Combines a traditional amortization schedule (Price / SAC) with
@@ -125,7 +122,7 @@ class BillingCycleLoan:
         self.cashflow = self._build_initial_cashflow()
 
     # ------------------------------------------------------------------
-    # Date derivation
+    # Date derivation (BCL-specific)
     # ------------------------------------------------------------------
 
     def _derive_closing_dates(self) -> List[datetime]:
@@ -171,13 +168,7 @@ class BillingCycleLoan:
         return all_dates[: self.num_installments]
 
     def _derive_due_dates(self) -> List[date]:
-        """Derive payment due dates from the billing cycle.
-
-        When the billing cycle has explicit due dates, use a wide
-        enough end boundary so that due dates falling after the last
-        closing date (the normal case -- due = close + offset) are
-        not accidentally excluded.
-        """
+        """Derive payment due dates from the billing cycle."""
         from dateutil.relativedelta import relativedelta
 
         last_closing = self._closing_dates[-1] if self._closing_dates else self.start_date
@@ -194,7 +185,7 @@ class BillingCycleLoan:
         return list(self._closing_dates)
 
     # ------------------------------------------------------------------
-    # CashFlow construction
+    # Abstract hook implementations
     # ------------------------------------------------------------------
 
     def _build_initial_cashflow(self) -> CashFlow:
@@ -240,10 +231,6 @@ class BillingCycleLoan:
 
         return CashFlow(items)
 
-    # ------------------------------------------------------------------
-    # Payment recording
-    # ------------------------------------------------------------------
-
     @tz_aware
     def record_payment(
         self,
@@ -255,19 +242,7 @@ class BillingCycleLoan:
         waive_mora: bool = False,
         discount: Optional[Money] = None,
     ) -> Settlement:
-        """Record a payment and return the derived settlement.
-
-        Args:
-            amount: Payment amount (must be positive).
-            payment_date: When the money moved.
-            interest_date: Cutoff for interest accrual.  Defaults to
-                *payment_date*.
-            description: Optional description.
-            waive_fines: If True, forgive accumulated fines.
-            waive_mora: If True, forgive accrued mora interest.
-            discount: Flat amount to forgive from obligations before
-                allocating the payment.
-        """
+        """Record a payment and return the derived settlement."""
         if amount.is_negative() or amount.is_zero():
             raise ValueError("Payment amount must be positive")
         if discount is not None and discount.is_negative():
@@ -292,86 +267,6 @@ class BillingCycleLoan:
 
         return self.settlements[-1]
 
-    def pay_installment(
-        self,
-        amount: Money,
-        description: Optional[str] = None,
-        waive_fines: bool = False,
-        waive_mora: bool = False,
-        discount: Optional[Money] = None,
-    ) -> Settlement:
-        """Pay the next installment.
-
-        Interest accrual depends on timing:
-
-        - Early / on-time: accrues up to the due date.
-        - Late: accrues up to ``now()`` (mora kicks in).
-
-        When all installments are already paid the payment is recorded
-        as an overpayment.
-
-        After recording the payment, if the principal balance drifts from
-        the schedule's expected ending balance by a small amount (within
-        ``payment_tolerance``), a tolerance adjustment CashFlowItem is
-        added to the cashflow to prevent rounding drift from compounding.
-
-        Args:
-            amount: Payment amount (must be positive).
-            description: Optional description.
-            waive_fines: If True, forgive accumulated fines.
-            waive_mora: If True, forgive accrued mora interest.
-            discount: Flat amount to forgive from obligations.
-        """
-        payment_date = self.now()
-
-        if self._covered_due_date_count() >= len(self.due_dates):
-            warnings.warn(
-                f"All installments already paid. Recording {amount} as overpayment.",
-                stacklevel=2,
-            )
-            return self.record_payment(
-                amount,
-                payment_date=payment_date,
-                interest_date=payment_date,
-                description=description or "Overpayment",
-                waive_fines=waive_fines,
-                waive_mora=waive_mora,
-                discount=discount,
-            )
-
-        next_due = self._next_unpaid_due_date()
-        interest_date = max(payment_date, self._time_ctx.to_datetime(next_due))
-        settlement = self.record_payment(
-            amount,
-            payment_date=payment_date,
-            interest_date=interest_date,
-            description=description,
-            waive_fines=waive_fines,
-            waive_mora=waive_mora,
-            discount=discount,
-        )
-
-        schedule = self.get_original_schedule()
-        for entry in schedule:
-            if entry.due_date == next_due:
-                apply_tolerance_adjustment(
-                    self.cashflow,
-                    entry,
-                    settlement,
-                    payment_date,
-                    interest_date,
-                    self.payment_tolerance,
-                    len(self.due_dates),
-                    self._time_ctx,
-                )
-                break
-
-        return settlement
-
-    # ------------------------------------------------------------------
-    # Derived state
-    # ------------------------------------------------------------------
-
     def _compute_state(self) -> LoanState:
         """Run the forward pass with per-cycle mora rate resolution."""
         return compute_state(
@@ -392,64 +287,7 @@ class BillingCycleLoan:
             calendar=self.working_day_calendar,
         )
 
-    def _payment_entries(self) -> list:
-        """Payment CashFlowEntry objects, sorted by datetime."""
-        entries = [e for e in self.cashflow.items() if "payment" in e.category]
-        return sorted(entries, key=lambda e: e.datetime)
-
-    # ------------------------------------------------------------------
-    # Settlements and installments
-    # ------------------------------------------------------------------
-
-    @property
-    def settlements(self) -> List[Settlement]:
-        """All settlements (derived from CashFlow)."""
-        return self._compute_state().settlements
-
-    @property
-    def installments(self) -> List[Installment]:
-        """Installment view (derived from CashFlow)."""
-        state = self._compute_state()
-        return build_installments(
-            self.get_original_schedule(),
-            state.settlements,
-            state.fines_applied,
-            state.principal_balance,
-            self.now(),
-            self._interest,
-            state.last_accrual_end,
-            tz=self._time_ctx.tz,
-            calendar=self.working_day_calendar,
-            grace_period_days=self.grace_period_days,
-        )
-
-    @property
-    def statements(self) -> List[BillingCycleLoanStatement]:
-        """Billing-period statements (one per cycle)."""
-        state = self._compute_state()
-        return build_statements(
-            self.get_original_schedule(),
-            self.due_dates,
-            self._closing_dates,
-            self.billing_cycle,
-            state.settlements,
-            state.fines_applied,
-            self.principal,
-            self.mora_interest_rate,
-            tz=self._time_ctx.tz,
-            mora_rate_resolver=self.mora_rate_resolver,
-        )
-
-    # ------------------------------------------------------------------
-    # Balance properties
-    # ------------------------------------------------------------------
-
-    @property
-    def principal_balance(self) -> Money:
-        """Outstanding principal."""
-        return self._compute_state().principal_balance
-
-    def _accrued_interest_components(self) -> tuple:
+    def _accrued_interest_components(self) -> Tuple[Money, Money]:
         """Return (regular, mora) accrued since last payment."""
         state = self._compute_state()
         days = (self._time_ctx.to_date(self.now()) - self._time_ctx.to_date(state.last_accrual_end)).days
@@ -482,41 +320,12 @@ class BillingCycleLoan:
         return Money.zero(), Money.zero()
 
     @property
-    def interest_balance(self) -> Money:
-        """Regular accrued interest since last payment."""
-        return self._accrued_interest_components()[0]
-
-    @property
-    def mora_interest_balance(self) -> Money:
-        """Mora accrued interest since last payment."""
-        return self._accrued_interest_components()[1]
-
-    @property
-    def fine_balance(self) -> Money:
-        """Unpaid fines."""
-        state = self._compute_state()
-        total_fines = (
-            Money(sum(f.raw_amount for f in state.fines_applied.values())) if state.fines_applied else Money.zero()
-        )
-        outstanding = total_fines - state.fines_paid_total
-        return outstanding if outstanding.is_positive() else Money.zero()
-
-    @property
-    def current_balance(self) -> Money:
-        """Total outstanding balance."""
-        return self.principal_balance + self.interest_balance + self.mora_interest_balance + self.fine_balance
-
-    @property
     def settlement_balance(self) -> Money:
         """Amount needed to cover the next installment via ``pay_installment``.
 
         Computes fines + mora + interest (accrued to ``max(now, next_due)``)
         + the next installment's scheduled principal.  Uses the per-cycle
         mora rate when a resolver is configured.
-
-        For single-installment or last-installment loans this equals the
-        full payoff amount.  For multi-installment loans with more than one
-        installment remaining, this is less than ``current_balance``.
         """
         state = self._compute_state()
         if not state.principal_balance.is_positive():
@@ -564,174 +373,35 @@ class BillingCycleLoan:
 
         return fine + mora + regular + next_entry.principal_payment
 
-    @property
-    def is_paid_off(self) -> bool:
-        """Whether the loan is fully paid off."""
-        if self.current_balance.is_zero() or self.current_balance.is_negative():
-            return True
-        return self._all_installments_covered()
-
-    def _all_installments_covered(self) -> bool:
-        """True when every installment has at least one fully-covered allocation.
-
-        Handles the case where schedule divergence leaves a small
-        positive balance but all per-installment allocations pass the
-        tolerance-based ``is_fully_covered`` check.
-
-        Guarded by a residual cap proportional to the number of
-        installments so a large balance is never silently accepted.
-        """
-        installments = self.installments
-        if not installments:
-            return False
-        n = len(self.due_dates)
-        max_residual = self.payment_tolerance * n * n
-        if self.current_balance > max_residual:
-            return False
-        return all(any(a.is_fully_covered for a in inst.allocations) for inst in installments)
-
-    @property
-    def overpaid(self) -> Money:
-        """Total amount paid beyond obligations."""
-        return self._compute_state().overpaid
-
     # ------------------------------------------------------------------
-    # Fine-related
+    # BCL-specific: late-check alias
     # ------------------------------------------------------------------
-
-    @property
-    def fines_applied(self) -> Dict[date, Money]:
-        """Fine amounts applied per due date."""
-        return self._compute_state().fines_applied
-
-    @property
-    def total_fines(self) -> Money:
-        """Total fines applied."""
-        fines = self.fines_applied
-        if not fines:
-            return Money.zero()
-        return Money(sum(f.raw_amount for f in fines.values()))
 
     @tz_aware
     def is_late(self, due_date: date, as_of_date: Optional[datetime] = None) -> bool:
         """Check if a payment is late considering the grace period."""
-        check = as_of_date if as_of_date is not None else self.now()
-        return is_payment_late(due_date, self.grace_period_days, check, self._time_ctx.tz, self.working_day_calendar)
-
-    def _on_warp(self, target_date: datetime) -> None:
-        """Hook called by Warp after overriding TimeContext."""
-        self._fine_observation_dates.append(target_date)
-
-    def calculate_late_fines(self, as_of_date: Optional[datetime] = None) -> Money:
-        """Compute and record fine observations as of a date.
-
-        Returns the amount of NEW fines applied.
-        """
-        as_of = as_of_date if as_of_date is not None else self.now()
-        old_total = self.total_fines
-        self._fine_observation_dates.append(as_of)
-        new_total = self.total_fines
-        return new_total - old_total
+        return self.is_payment_late(due_date, as_of_date)
 
     # ------------------------------------------------------------------
-    # Payment info
+    # BCL-specific: Statements
     # ------------------------------------------------------------------
 
     @property
-    def last_payment_date(self) -> datetime:
-        """Date of the last payment, or disbursement date if none."""
-        return self._compute_state().last_payment_date
-
-    def now(self) -> datetime:
-        """Current datetime (Warp-aware)."""
-        return self._time_ctx.now()
-
-    def days_since_last_payment(self) -> int:
-        """Days since the last payment."""
-        return (self._time_ctx.to_date(self.now()) - self._time_ctx.to_date(self.last_payment_date)).days
-
-    def _covered_due_date_count(self) -> int:
-        return covered_due_date_count(self.principal_balance, self.get_original_schedule())
-
-    def _next_unpaid_due_date(self) -> date:
-        covered = self._covered_due_date_count()
-        if covered >= len(self.due_dates):
-            raise ValueError("All due dates have been paid")
-        return self.due_dates[covered]
-
-    # ------------------------------------------------------------------
-    # Schedule
-    # ------------------------------------------------------------------
-
-    def get_original_schedule(self) -> PaymentSchedule:
-        """The original amortization schedule (static, ignores payments)."""
-        return self.scheduler.generate_schedule(
-            self.principal,
-            self.interest_rate,
-            self.due_dates,
-            self.disbursement_date,
-            self._time_ctx.tz,
-        )
-
-    def get_amortization_schedule(self) -> PaymentSchedule:
-        """Current schedule: recorded past entries + projected future."""
+    def statements(self) -> List[BillingCycleLoanStatement]:
+        """Billing-period statements (one per cycle)."""
         state = self._compute_state()
-        if not state.settlements:
-            return self.get_original_schedule()
-
-        actual_entries: List[PaymentScheduleEntry] = []
-        prev_balance = self.principal
-        prev_date = self.disbursement_date
-
-        for i, s in enumerate(state.settlements):
-            days = (self._time_ctx.to_date(s.payment_date) - self._time_ctx.to_date(prev_date)).days
-            actual_entries.append(
-                PaymentScheduleEntry(
-                    payment_number=i + 1,
-                    due_date=self._time_ctx.to_date(s.payment_date),
-                    days_in_period=days,
-                    beginning_balance=prev_balance,
-                    payment_amount=s.interest_paid + s.mora_paid + s.principal_paid,
-                    principal_payment=s.principal_paid,
-                    interest_payment=s.interest_paid + s.mora_paid,
-                    ending_balance=s.remaining_balance,
-                )
-            )
-            prev_balance = s.remaining_balance
-            prev_date = s.payment_date
-
-        covered = covered_due_date_count(state.principal_balance, self.get_original_schedule())
-        remaining_due_dates = self.due_dates[covered:]
-        if not remaining_due_dates:
-            return PaymentSchedule(entries=actual_entries)
-
-        if state.principal_balance.is_zero() or state.principal_balance.is_negative():
-            return PaymentSchedule(entries=actual_entries)
-
-        projected_schedule = self.scheduler.generate_schedule(
-            state.principal_balance,
-            self.interest_rate,
-            remaining_due_dates,
-            state.last_payment_date,
-            self._time_ctx.tz,
+        return build_statements(
+            self.get_original_schedule(),
+            self.due_dates,
+            self._closing_dates,
+            self.billing_cycle,
+            state.settlements,
+            state.fines_applied,
+            self.principal,
+            self.mora_interest_rate,
+            tz=self._time_ctx.tz,
+            mora_rate_resolver=self.mora_rate_resolver,
         )
-
-        projected_entries: List[PaymentScheduleEntry] = []
-        for entry in projected_schedule:
-            projected_entries.append(
-                PaymentScheduleEntry(
-                    payment_number=len(actual_entries) + entry.payment_number,
-                    due_date=entry.due_date,
-                    days_in_period=entry.days_in_period,
-                    beginning_balance=entry.beginning_balance,
-                    payment_amount=entry.payment_amount,
-                    principal_payment=entry.principal_payment,
-                    interest_payment=entry.interest_payment,
-                    ending_balance=entry.ending_balance,
-                )
-            )
-
-        return PaymentSchedule(entries=actual_entries + projected_entries)
 
     # ------------------------------------------------------------------
     # Dunder
